+++ src/modules/newsFeed/NewsFeedSlice.ts
@@ -2,6 +2,7 @@ import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import axios from 'axios';
 
 import { PaginatedApiResponse } from '@/modules/common/CommonTypes';
+import { RootState } from '@/store';
 
 import * as NewsFeedApi from './NewsFeedApi';
 import { NewsItemType } from './NewsFeedTypes';
@@ -10,9 +11,13 @@ interface NewsFeedState {
     items: NewsItemType[];
     count: number;
     currentPage: number;
-    isLastPage: boolean;
     loading: boolean;
     error: string;
+
+    // Number of unseen news since the last time user opened the news feed
+    actualizedCount: number;
+    actualizedCountLoading: boolean;
+    actualizedCountError: string;
 }
 
 const initialState: NewsFeedState = {
@@ -20,22 +25,59 @@ const initialState: NewsFeedState = {
     count: 0,
     loading: true,
     currentPage: 1,
-    isLastPage: false,
     error: '',
+
+    actualizedCount: 0,
+    actualizedCountLoading: true,
+    actualizedCountError: '',
 };
 
 export const NEWS_FEED_PAGE_SIZE = 10;
 
+type listNewsFeedArgs = {
+    firstPage: boolean;
+    timestamp_before?: string;
+};
 export const listNewsFeed = createAsyncThunk<
     PaginatedApiResponse<NewsItemType>,
-    { firstPage: boolean },
+    listNewsFeedArgs,
     { rejectValue: string }
->('newsFeed/list', async ({ firstPage }, thunkAPI) => {
-    const state = thunkAPI.getState() as { newsFeed: NewsFeedState };
-    const page = firstPage ? 1 : state.newsFeed.currentPage;
+>(
+    'newsFeed/list',
+    async ({ firstPage, timestamp_before }: listNewsFeedArgs, thunkAPI) => {
+        const state = thunkAPI.getState() as RootState;
+        const page = firstPage ? 1 : state.newsFeed.currentPage;
+
+        try {
+            const response = await NewsFeedApi.listNewsFeed(
+                {
+                    page,
+                    pageSize: NEWS_FEED_PAGE_SIZE,
+                    timestamp_before,
+                },
+                {
+                    signal: thunkAPI.signal,
+                }
+            );
+            return response.data;
+        } catch (error) {
+            if (axios.isAxiosError(error)) {
+                return thunkAPI.rejectWithValue(error.response?.data);
+            } else {
+                throw error;
+            }
+        }
+    }
+);
+
+export const retrieveActualizedCount = createAsyncThunk<
+    PaginatedApiResponse<NewsItemType>,
+    { timestamp_after: string },
+    { rejectValue: string }
+>('newsFeed/actualizedCount', async ({ timestamp_after }, thunkAPI) => {
     try {
         const response = await NewsFeedApi.listNewsFeed(
-            { page: page, pageSize: NEWS_FEED_PAGE_SIZE },
+            { pageSize: 0, timestamp_after },
             {
                 signal: thunkAPI.signal,
             }
@@ -73,7 +115,6 @@ const newsFeedSlice = createSlice({
                     state.items = [...state.items, ...payload.results];
                 }
                 state.count = payload.count;
-                state.isLastPage = !payload.next;
                 state.loading = false;
                 state.error = '';
             })
@@ -82,6 +123,23 @@ const newsFeedSlice = createSlice({
                     state.loading = false;
                     state.error = payload || 'Unknown error';
                 }
+                state.currentPage = state.currentPage - 1;
+            })
+            .addCase(retrieveActualizedCount.pending, (state) => {
+                state.actualizedCountLoading = true;
+                state.actualizedCountError = '';
+            })
+            .addCase(
+                retrieveActualizedCount.fulfilled,
+                (state, { payload }) => {
+                    state.actualizedCount = payload.count;
+                    state.actualizedCountLoading = false;
+                    state.actualizedCountError = '';
+                }
+            )
+            .addCase(retrieveActualizedCount.rejected, (state, { payload }) => {
+                state.actualizedCountLoading = false;
+                state.actualizedCountError = payload || 'Unknown error';
             });
     },
 });