+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Newsfeed notifications (#588)
+
 ### Fixed
 
 -   Removed date filters glitch (#585)